refector: 로또 개수 구하는 로직 클래스 분리
