+++ src/main/java/lotto/controller/GameController.java
@@ -1,10 +1,9 @@
 package lotto.controller;
 
 import lotto.domain.Player;
+import lotto.service.LottoCountGenerator;
 import lotto.service.PlayerService;
 import lotto.service.WinningNumberGenerator;
-import lotto.util.ErrorMessage;
-import lotto.util.LottoValues;
 
 public class GameController {
 
@@ -12,47 +11,27 @@ public class GameController {
     private PlayerService playerService;
     private WinningNumberGenerator winningNumberGenerator;
     private OutputController outputController;
+    private LottoCountGenerator lottoCountGenerator;
     private int lottoCount;
     private int customerPrice;
 
     public GameController(Player player, PlayerService playerService, WinningNumberGenerator winningNumberGenerator,
-                          OutputController outputController) {
+                          OutputController outputController, LottoCountGenerator lottoCountGenerator) {
         this.player = player;
         this.playerService = playerService;
         this.winningNumberGenerator = winningNumberGenerator;
         this.outputController = outputController;
+        this.lottoCountGenerator = lottoCountGenerator;
     }
 
     public void start() {
-        setLottoCount();
+        customerPrice = lottoCountGenerator.getCustomerPrice();
+        lottoCount = lottoCountGenerator.getLottoCount(customerPrice);
+
         playerService.buy(lottoCount);
         outputController.printLottoPapers(player.getLottoPapers(), lottoCount);
 
         playerService.check(winningNumberGenerator.generate());
         outputController.printWinningStatistics(customerPrice);
-    }
-
-    private void setLottoCount() {
-        while(true){
-            try{
-                customerPrice = InputController.inputPrice();
-                lottoCount = parseLottoCount(customerPrice);
-                break;
-            }catch(IllegalArgumentException e){
-                System.out.println(ErrorMessage.ERROR_INPUT_PRICE_MESSAGE);
-                System.out.println();
-            }
-        }
-    }
-
-    private int parseLottoCount(int customerPrice) throws IllegalArgumentException{
-        if (canChangeCountByLottoPrice(customerPrice)) {
-            throw new IllegalArgumentException();
-        }
-        return customerPrice / LottoValues.LOTTO_PRICE;
-    }
-
-    private static boolean canChangeCountByLottoPrice(int customerPrice) {
-        return (customerPrice % LottoValues.LOTTO_PRICE != 0) || (customerPrice == 0);
     }
 }
